web/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pyyaml>=6.0
pydantic>=2.0
aiofiles>=23.0.0
